+++ src/memory.h
@@ -21,9 +21,9 @@ class Memory {
     Console* console;
     uint8_t mem_array[0X10000]; // the size of the array full address space for
                                 // the banana
-    bool isReadable(uint16_t address) const;
-    bool isWritable(uint16_t address) const;
-    bool isExecutable(uint16_t address) const;
+    bool isReadable(uint32_t address) const;
+    bool isWritable(uint32_t address) const;
+    bool isExecutable(uint32_t address) const;
 
   public:
     Memory(Console* console);
